standalone-skills/requirements.txt
PyMuPDF>=1.23.0
PyYAML>=6.0
requests>=2.31.0
